sign up dialog offically pushes user info into DB!
